# openapi.yaml
openapi: 3.0.3
info:
  title: ZTN_SIM IAMaaS API
  version: "1.0.0"
  description: |
    Versioned REST contract for IAMaaS tenants and core platform operations. The schema documents
    authentication, token issuance, trust-score inspection, policy management, audit/log retrieval,
    and supporting wallet/transaction flows. It is intended to be consumed by the auth-gateway,
    policy decision point, external HMS integrations, and telecom or mobile-money pilots.
servers:
  - url: https://localhost.localdomain.com
    description: Local HTTPS (Podman + nginx)
  - url: http://localhost.localdomain.com
    description: Local HTTP (if configured)
security:
  - BearerAuth: []
    ApiKeyAuth: []
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create a new user record
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
  /user:
    get:
      tags: [Users]
      summary: Retrieve authenticated user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Users]
      summary: Update authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
    delete:
      tags: [Users]
      summary: Delete authenticated user
      security:
        - BearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
  /user/profile:
    get:
      tags: [Users]
      summary: Retrieve concise profile information
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /user/request_deletion:
    post:
      tags: [Users]
      summary: Request account deletion
      security:
        - BearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
  /user-info/{mobileNumber}:
    get:
      tags: [Users]
      summary: Fetch user info by mobile number
      security:
        - BearerAuth: []
      parameters:
        - name: mobileNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLookup'
        '404': { $ref: '#/components/responses/NotFound' }
  /wallets:
    get:
      tags: [Wallets]
      summary: List wallets or retrieve current balance
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wallet'
    put:
      tags: [Wallets]
      summary: Update wallet details
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletUpdateRequest'
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
    delete:
      tags: [Wallets]
      summary: Remove a wallet association
      security:
        - BearerAuth: []
      responses:
        '204': { description: Wallet removed }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MessageResponse:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Message:
      type: object
      properties:
        message:
          type: string
...
    UserCreateRequest:
      type: object
      required: [mobile_number, first_name]
      properties:
        mobile_number:
          type: string
        first_name:
          type: string
        country:
          type: string
        identity_verified:
          type: boolean
    UserUpdateRequest:
      type: object
      properties:
        first_name:
          type: string
        country:
          type: string
        identity_verified:
          type: boolean
        password:
          type: string
          format: password
    User:
      type: object
      properties:
        id:
          type: integer
        mobile_number:
          type: string
        first_name:
          type: string
        email:
          type: string
        country:
          type: string
        identity_verified:
          type: boolean
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        mobile_number:
          type: string
        first_name:
          type: string
        country:
          type: string
        identity_verified:
          type: boolean
    UserLookup:
      type: object
      properties:
        id:
          type: integer
        mobile_number:
          type: string
        name:
          type: string
    Wallet:
      type: object
      properties:
        id:
          type: integer
        balance:
          type: number
          format: float
        currency:
          type: string
    WalletUpdateRequest:
      type: object
      properties:
        balance:
          type: number
          format: float
        currency:
          type: string
...